requirements.txt
argparse==1.4.0
beautifulsoup4==4.4.1
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.7
django-anymail==0.3.1
django-bootstrap3==7.0.1
django-htmlmin==0.9.1
django-modelcluster==2.0
django-taggit==0.18.3
django-taggit-templatetags2==1.5.1
django-toolbelt==0.0.1
django-treebeard==4.0.1
djangorestframework==3.3.3
gunicorn==19.6.0
html5lib==0.9999999
Pillow==3.2.0
psycopg2==2.6.1
pytz==2016.4
requests==2.10.0
six==1.10.0
static3==0.7.0
Unidecode==0.4.19
wagtail==1.5.2
whitenoise==3.2
Willow==0.3.1
